requirements.txt
streamlit>=1.28.0
pdfplumber>=0.11.0
python-docx>=1.1.0
openai>=1.40.0
python-dotenv>=1.0.0
# Scanned PDF: OpenAI vision (pymupdf) + optional Tesseract
pymupdf>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
